+++ Plugins/Appboy/Models/ClickAction.cs
@@ -1,0 +1,6 @@
+namespace Appboy.Models {
+  public enum ClickAction {
+    NEWS_FEED, URI, NONE
+  }
+}
+