+++ Plugins/Appboy/Models/DismissType.cs
@@ -1,0 +1,6 @@
+namespace Appboy.Models {
+  public enum DismissType {
+    AUTO_DISMISS, SWIPE
+  }
+}
+