+++ Plugins/Appboy/Models/Gender.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+
+namespace Appboy.Models {
+  public enum Gender {
+    Male = 1,
+    Female = 2
+  }
+}