+++ Plugins/Appboy/models/Gender.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-
-namespace Appboy {
-  public enum Gender {
-    Male = 1,
-    Female = 2
-  }
-}