+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2
+* Updated the plugin libraries to use the Appboy Android SDK v1.4.3 and Appboy iOS SDK v2.8 (without Facebook iOS SDK Support) libraries
+* Added SlideFrom, ClickAction, DismissType and Uri to Slideup; added logging slideup impressions and clicks.
+* Exposed the card models from Appboy to Unity; added methods for requesting feed from Appboy server or cache; added logging impressions and clicks
+* In Android SDK, changed the device identifier from the device persistent ANDROID_ID to a non device persistent identifier for compliance with the new Google Play Terms of Service.
+
 ## 1.1
 * Updating the plugin libraries to the Appboy Android SDK v1.2.1 and Appboy iOS SDK v2.3.1 (without Facebook iOS SDK Support) libraries 
 * Makes compatible with Prime31 plugins